+++ src/LDL/Router/Http/Route/HttpRouteInterface.php
@@ -4,6 +4,8 @@ declare(strict_types=1);
 
 namespace LDL\Router\Http\Route;
 
+use LDL\Http\Core\Response\ResponseInterface;
+use LDL\Router\Core\Route\Path\Result\RoutePathMatchingResultInterface;
 use LDL\Router\Core\Route\RouteInterface;
 use LDL\Router\Http\Collection\HttpMethodCollection;
 use LDL\Router\Http\Response\Encoder\HttpResponseEncoderInterface;
@@ -15,4 +17,6 @@ interface HttpRouteInterface extends RouteInterface
     public function getMethods(): HttpMethodCollection;
 
     public function getResponseEncoder(): HttpResponseEncoderInterface;
+
+    public function dispatch(RoutePathMatchingResultInterface $path, ResponseInterface $response): void;
 }